import protocol part

+++ tasks/broadcast_task.go
@@ -40,7 +40,7 @@ func (t *BroadcastTask) Process(tl types.Timeline) {
 
 	// calculate bandwidth usage time.
 	totalMsgSize := t.msg.Size * int64(len(destinations))
-	bandwidthTimeUsage := totalMsgSize / int64(n.BandwidthInMillisecond())
+	bandwidthTimeUsage := totalMsgSize / n.BandwidthInMillisecond()
 
 	for _, dest := range destinations {
 		msg := t.msg
